Prevent solution data interfering with submission's
Both the solution and submission were being wrapped, and
given the same `dataPath` to write their files to. This
would generally result in `exercise.wrapData` coming
from the solution instead of the submission, though
sometimes the solution would finish early and it would
run as expected.

Also corrected a check in exec-wrap.js to properly
identify whether the solution should be wrapped.